pyproject.toml
[project]
name = "sherlook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "pyaml>=25.7.0",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]
